# Optimise ReadString()