… protocol/binary/bench_after.out
… protocol/binary/bench_before.out
… protocol/binary/cpu_after.out
… protocol/binary/cpu_before.out