Optimise ReadString() CPU by ~30%
```
name          old time/op  new time/op  delta
ReadString    90.4ns ± 4%  63.1ns ± 5%  -30.13%  (p=0.008 n=5+5)
ReadString-2   103ns ±10%    68ns ± 7%  -33.31%  (p=0.008 n=5+5)
ReadString-4   109ns ± 7%    79ns ±29%  -27.92%  (p=0.008 n=5+5)
```